Fix initial setup flow.
As procd would only retry 5 times, the Concentratord, MQTT forwarder and
ChirpStack would not be restarted after initial setup. This changes the
procd respawn to unilimited times.

As well this always forces the channel_plan write. As the channel_plan
write callback must always be triggered, to make sure the MQTT
Forwarder and ChirpStack will be re-configured properly.